Fix AIService warnings
